+++ activities/Flip.activity/js/activity.js
@@ -51,7 +51,7 @@ function runactivity(act,doc,colors,env,datastore,sizepalette,tutorial){
 		window.addEventListener('resize', resizeCanvas, false);
 		function resizeCanvas() {
 			canvas.width = window.innerWidth;
-			canvas.height = window.innerHeight;
+			canvas.height = window.innerHeight-(+document.getElementById("main-toolbar").style.opacity ? 55 : 0);
 			g.initialiseFromArray();
 		}
 		var solveButton = doc.getElementById("solve-button");
@@ -72,6 +72,20 @@ function runactivity(act,doc,colors,env,datastore,sizepalette,tutorial){
 		window.addEventListener('activityStop', function (eve) {
 			eve.preventDefault();
 			g.stop();
+		});
+
+		// Full screen
+		document.getElementById("fullscreen-button").addEventListener('click', function() {
+			document.getElementById("main-toolbar").style.opacity = 0;
+			document.getElementById("canvas").style.top = "0px";
+			document.getElementById("unfullscreen-button").style.visibility = "visible";
+			resizeCanvas();
+		});
+		document.getElementById("unfullscreen-button").addEventListener('click', function() {
+			document.getElementById("main-toolbar").style.opacity = 1;
+			document.getElementById("canvas").style.top = "55px";
+			document.getElementById("unfullscreen-button").style.visibility = "hidden";
+			resizeCanvas();
 		});
 	}
 	init();