+++ CHANGELOG.md
@@ -45,6 +45,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for no signup mode
 - Add a tutorial to the Flip activity #442
 - Add a tutorial to the PhysicsJS activity #428
+- Add a fullscreen button in Physics activity #410
+- Add a fullscren button in GameOfLife activity #407
+- Add a fullscreen button in the Flip activity #514
 
 ### Changed
 - Double size of Record images in WebApp and electron App
@@ -65,6 +68,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Image insertion don't work in shared mode in Paint activity #384
 - Close the presence palette when the activity is shared #398
 - Board size in Flip activity is not optimized #405
+- Make responsive the toolbar in Write activity #515
+- Make responsive the toolbar in Physics activity #516
 
 
 ## [1.2.0] - 2019-09-26